Slim down samsung_ag/trogdor-framework image
Used centurylink/golang-builder to build the hello binary.  I don't like
checking binaries in, but our CI process currently assumes the
Dockerfile is authoritative.

The hello binary built with:
  docker run --rm -v $(pwd):/src centurylink/golang-builder

This could be scripted away or the golang-builder's docker.sock volume
could be used to automate away the need to check in a binary