Migrate from heaspter,influxdb-grafana to cluster-montoring
Merge the heapster, influxdb-grafana services into a cluster-monitoring
service based on the cluster/addons/cluster-monitoring/influxdb service
from kubernetes.  Bump versions as appropriate.

NB: We're going to lose the ability to demo loadtest service writing into
influxdb until we add back the capability to precreate a database for
that, and put the influxdb service in a namespace that loadtest can also
see.

Some customizations:
- use NodePort service types where possible
- add back the custom kubernetes.io/color label used by podpincher